help context awareness
Increase context awareness of help bar. Intelligently display available
commands based on whether or not the current process is running, and if
it has a terminal available.

This commit has a bug which should be resolved before merging. The
terminal manager associated with the first process in the list always
returns None from get_terminal.