Refactor keybindings
Instead of having multiple components with registered keybindings, the
TUIController will have sole responsibility for registering appropriate
keybindings with the Application instance. Registered keybindings will
change as the state of the app changes.

This simplifies the help component as well. It no longer needs to worry
about if a help label should be displayed based on which element has
focus, and the state of the app. It simply needs to display labels for
all registered keybindings at any given time.